+++ hal/avr/make/library.mk
