+++ hal/avr/mocks/attinyx5/io.c
