+++ hal/avr/tests/attinyx5/test_pin_attinyx5.cpp
@@ -1,0 +1,22 @@
+/**
+ * @file      test_pin_attinyx5.c
+ * @brief     This file contains tests for ATtinyx5 series pin driver.
+ * @author    Liam Bucci <liam.bucci@gmail.com>
+ * @date      2017-09-22
+ * @copyright Copyright (c) 2017 Liam Bucci. See included LICENSE file.
+ */
+
+#include <stdio.h>
+
+#include "unity.h"
+
+#include "junk/hal/attinyx5/pin.h"
+
+using namespace junk::hal;
+
+int main(int argc, char** argv)
+{
+    UNITY_BEGIN();
+
+    return UNITY_END();
+}
